requirements.txt
tavily-python
google-generativeai
crawl4ai
resend
markdown
python-dotenv
python-dateutil
beautifulsoup4
